♻️Refactor: 청중 분포 확률로 변경, 세션 종료 발표자 JWT 검증 삭제, 새로 방 입장 시 현재 페이지로 이동

+++ src/main/java/line4thon/boini/presenter/page/dto/response/SlideAudienceCountResponse.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class SlideAudienceCountResponse {
-    private int frontCount;
-    private int currentCount;
-    private int backCount;
+    private Long frontCount;
+    private Long currentCount;
+    private Long backCount;
 }
